lib: data-stack - fix pointer arithmetic compiler warning Clang's -fsanitize=unsigned-integer-overflow barfs as follows:
data-stack.c:477:29: runtime error: negation of 8 cannot be represented in type 'unsigned long'
data-stack.c:495:15: runtime error: negation of 8 cannot be represented in type 'unsigned long'

Which is of course complete bollocks. There is no 8, there's only an 8ul, and
the negation of 8ul is 0xfffffff8ul (or a wider equivalent). That's the law.

However, the pointer arithmetic which follows the negation, whilst almost
certainly working in practice, is probably bogus, so just make the thing
signed before the negation, and both problems go away.

Reported-by: Teemu Huovila <teemu.huovila@dovecot.fi>

+++ src/lib/data-stack.c
@@ -474,7 +474,7 @@ bool t_try_realloc(void *mem, size_t size)
 		alloc_growth = (new_alloc_size - last_alloc_size);
 #ifdef DEBUG
 		size_t old_raw_size; /* sorry, non-C99 users - add braces if you need them */
-		old_raw_size = *(size_t *)PTR_OFFSET(mem, -MEM_ALIGN(sizeof(size_t)));
+		old_raw_size = *(size_t *)PTR_OFFSET(mem, -(ptrdiff_t)MEM_ALIGN(sizeof(size_t)));
 		i_assert(ALLOC_SIZE(old_raw_size) == last_alloc_size);
 		/* Only check one byte for over-run, that catches most
 		   offenders who are likely to use t_try_realloc() */
@@ -492,7 +492,7 @@ bool t_try_realloc(void *mem, size_t size)
 			/* All reallocs are permanent by definition
 			   However, they don't count as a new allocation */
 			current_frame_block->alloc_bytes[frame_pos] += alloc_growth;
-			*(size_t *)PTR_OFFSET(mem, -MEM_ALIGN(sizeof(size_t))) = size;
+			*(size_t *)PTR_OFFSET(mem, -(ptrdiff_t)MEM_ALIGN(sizeof(size_t))) = size;
 			memset(PTR_OFFSET(mem, size), CLEAR_CHR,
 			       new_alloc_size - size - MEM_ALIGN(sizeof(size_t)));
 #endif
